+++ vendas-domain/src/main/java/br/com/vendas/domain/order/PriceTable.java
@@ -18,13 +18,13 @@ public class PriceTable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long ID;
+	private Integer ID;
 	
 	@Column(name="IDEMPRESA")
-	private Long organizationID;
+	private Integer organizationID;
 	
 	@Column(name="IDFILIAL")
-	private Long branchID;
+	private Integer branchID;
 	
 	@Column(name="IDTABELA")
 	private Integer priceTableID;
@@ -49,28 +49,28 @@ public class PriceTable {
 	private boolean excluded;
 					
 	
-	public Long getID() {
+	public Integer getID() {
 		return ID;
 	}
 
-	public void setID(Long iD) {
+	public void setID(Integer iD) {
 		ID = iD;
 	}
 
-	public Long getOrganizationID() {
+	public Integer getOrganizationID() {
 		return organizationID;
 	}
 
-	public void setOrganizationID(Long organizationID) {
+	public void setOrganizationID(Integer organizationID) {
 		this.organizationID = organizationID;
 	}
 
 	
-	public Long getBranchID() {
+	public Integer getBranchID() {
 		return branchID;
 	}
 
-	public void setBranchID(Long branchID) {
+	public void setBranchID(Integer branchID) {
 		this.branchID = branchID;
 	}
